some small patches
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@745 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/BufferView_pimpl.C
@@ -469,7 +469,7 @@ int BufferView::Pimpl::scrollDown(long time)
 	if (!screen) return 0;
    
 	double value= workarea->getScrollbarValue();
-	pair<double, double> p = workarea->getScrollbarBounds();
+	pair<float, float> p = workarea->getScrollbarBounds();
 	double max = p.second;
 	
 	if (value == max) return 0;
